move files to new dir; wrapper multi_handler to Channel
