--- net/grpc/web/echo_example/composer.json
@@ -1,13 +1,0 @@
-{
-	"name": "grpc/grpc-demo",
-	"description": "gRPC example for PHP",
-	"require": {
-		"google/protobuf": "^v3.3.0"
-	},
-	"autoload": {
-		"psr-4": {
-			"": "./",
-			"Grpc\\": "../grpc/src/lib"
-		}
-	}
-}
