--- net/grpc/web/grpc/composer.json
@@ -1,24 +1,0 @@
-{
-  "name": "grpc/grpc",
-  "type": "library",
-  "description": "gRPC library for PHP",
-  "keywords": ["rpc"],
-  "homepage": "https://grpc.io",
-  "license": "Apache-2.0",
-  "version": "1.6.0",
-  "require": {
-    "php": ">=5.5.0"
-  },
-  "require-dev": {
-    "google/auth": "v0.9"
-  },
-  "suggest": {
-    "ext-protobuf": "For better performance, install the protobuf C extension.",
-    "google/protobuf": "To get started using grpc quickly, install the native protobuf library."
-  },
-  "autoload": {
-    "psr-4": {
-      "Grpc\\": "src/lib/"
-    }
-  }
-}
